Update pipeline references when resetting pipeline state
To reduce the potential memory usage from change
I7fa99cd83a857216321f8d946fd42abd9ec427a3, this updates the
pipeline references in the ChangeQueue instances underneath a
given PipelineState when that state is being reset after a
reconfiguration.

It also removes the pipeline reference from the QueueItem,
replacing it with a property that uses the pipeline reference
on the ChangeQueue instead.  This removes one extra place where
an incorrect reference could cause a memory leak.

With these two changes, there should be no more references to
outdated layouts/pipelines in the time period between a
reconfiguration and the subsequent pipeline refresh.

Change-Id: If1428a11277e8e1f03e2cc8f965f7db6ac79c308