Avoid unnecessary change dependency updates
Updating commit and topic dependencies incurres a cost as we query the
source system for the change details.

The current implementation will update the commit and topic dependencies
independent of whether or not the dependencies are already populated and
when they were last updated. This can lead to multiple updates for the
same change in a short amount of time e.g. when an event leads to a
change to be added to multiple pipelines or when a circular dependency
is enqueued.

Instead we can use the following conditions to determine if the
dependencies need to be refreshed:

1. when `updateCommitDependencies()` is called without an event (force a
   dependency refresh)
2. when the change's commit or topic dependencies were never updated
3. when the event ltime is greater than the last modified zxid of the
   change (dependencies could have change in the meantime)

Change-Id: I4fd6c0d4cf2839010ddf7105a7db12da06ef1074