Add tox v4 compatibility
This converts whitelist_externals to allowlist_externals for tox v4
compatibility. Additionally, we drop skipsdist as this seems to cause
usedevelop to be ignored. Reviewers should look at this closely to
ensure the right combo of skipsdist, usedevelop, and install_command are
used.

Update tools/pip.sh to check if zuul-manage-ansible is present before
running it. We do this because tox v4 has slightly changed the
dependency installation process and we were attempting to run this
command before it had been installed.

Depends-On: https://review.opendev.org/c/zuul/zuul-jobs/+/866943
Change-Id: I20897716de08776af35bf110bf8288da3cde4a61