Add tox v4 compatibility
This converts whitelist_externals to allowlist_externals for tox v4
compatibility. Additionally, we drop skipsdist as this seems to cause
usedevelop to be ignored. Reviewers should look at this closely to
ensure the right combo of skipsdist, usedevelop, and install_command are
used.

We remove basepython beacuse new tox seems to use this to determine what
version of python to create a venv for. This results in improperly
running python3.11 envs under python3.10 beacuse python3.10 is the
default on Ubuntu Jammy (it owns `python3`). This removal should be fine
as it was primariy there to force the use of python3 during the python2
to python3 transition. We're well beyond that.

Add a setuptools package find: option to setup.cfg which causes pep 517
builders (used when tox does build and sdist under the linters env) to
properly find our package contents. As far as I can tell via testing
with the `build` too this results in copying the same set of files that
`build` produced previously.

Update tools/pip.sh to check if zuul-manage-ansible is present before
running it. We do this because tox v4 has slightly changed the
dependency installation process and we were attempting to run this
command before it had been installed.

Depends-On: https://review.opendev.org/c/zuul/zuul-jobs/+/866943
Change-Id: I20897716de08776af35bf110bf8288da3cde4a61