Track object versions in the Buildset object
This further reduces the number of ZK object reads during pipeline
refreshes by tracking when builds and frozen jobs are updated.

During the phases of a build where we know no updates can occur,
we already avoid refreshing the Build and FrozenJob objects.
But, for example, while a build is running we have to continually
refresh it to see if it has completed.

We can avoid this by recording expected version information in ZK
and only refresh those objects if we know our local copy is out
of date.

We can store the latest ZK object version of FrozenJob and Build
objects on the Buildset.  On pipeline refresh, we currently
always refresh the buildset object, which means that when we
prepare to refresh the FrozenJob or Build objects underneath a
Buildset, we will have information about the latest versions of
those objects in ZK and can compare to the versions we currently
have in memory to decide if we need to refresh them.  This should
reduce the number of reads in a pipeline refresh by about 50%.
But it will cause more writes, in that we will update the
Buildset object each time we modify one of its children.  This
may affect pipeline processing times but the impact should be
very small.

We will use version numbers (rather than transaction ids) because they
are predictable, and updating the buildset first with the predicted
next version before updating the child avoids issues caused by a crash
between those two steps.

Since it is typical for many objects to be created at once, we do
optimize the case where the objects are initially created and we
avoid making an update to the BuildSet in that case so that we
don't repeatedly write the buildset object.

Change-Id: I3824af6149bf27c41a8d895fc682236bd0d91f6b