Add timer event directly to pipeline event queues
Timer events are always generated for a particular pipeline and tenant.
Because of that we can directly add the timer triggers to the matching
pipeline event queue instead of taking a detour via the tenant trigger
event queue.

This will also eliminate duplicate events in cases where a project is
part of multiple tenants. This happened since the scheduler added the
event to all tenants that have the project configured that the event
refers to.

Change-Id: I6b4b77a82aac6d2c3441ace9f64dc9b4a80c5856