Further avoid unnecessary change dependency updates
When adding a unit test for change I4fd6c0d4cf2839010ddf7105a7db12da06ef1074
I noticed that we were still querying the dependent change 4 times instead of
the expected 2.  This was due to an indentation error which caused all 3
query retry attempts to execute.

This change corrects that and adds a unit test that covers this as well as
the previous optimization.

Change-Id: I798d8d713b8303abcebc32d5f9ccad84bd4a28b0