Remove usage of MR diff_refs attribute
Since recently Gitlab updates the diff_refs attribute of a merge request
asynchronously. This is discussed here:

https://gitlab.com/gitlab-org/gitlab/-/issues/386562

Usage of that attribute is made difficult and this patch removes its
usage.

The change's `base_sha` attribute is not needed to be set for a change as if
not provided the base branch's tip SHA is used by Zuul internally. The
change's `commit_id` attribute is set via the event's `last_commit`
attribute.

Change-Id: Ia819291069e7ed1080bcd3916736b6a311706cce