Unvendor kazoo locks recipes
These recipes were vendored so that we could carry this fix locally:
  https://github.com/python-zk/kazoo/pull/650

It appears that this fix has been merged and included in kazoo>=2.9.0 so
we include that as the minimum version and drop the vendored file.

This also fixes the isSet() deprecation warning as upstream kazoo has
switched to is_set().

Change-Id: Ide48e9f949e083b658775b74db3856b118fc5d69