Refresh dependencies for changes in pipelines
Commit 549103881899751e84b411212f7820cde502b42b appears to have
had an error.  It attempted to only refresh changes if they are
in the pipeline or are dependencies of changes in the pipeline,
but it only did the latter, which has an adverse effect for the
GitHub driver.  We can change dependencies of GitHub changes
without creating new patchsets, which means that this error
(not updating dependencies of changes in the pipeline) would cause
Zuul not to eject a GitHub change from a pipeline when its
dependencies changed via a PR message update.

The error went unnoticed with Gerrit since dependency updates
happen with new patchsets, which themselves cause changes to be
ejected.

This change corrects the behavior.

Repetitive calls to updateCommitDependencies are also eliminated
by creating a set of changes to update across the whole pipeline.

Finally, two tests are added, one for Gerrit (which previously would
have passed) and one for GitHub (which previously would have failed).

Change-Id: I005820223471e8f6372ef70730f381918ec9c1af