Update build/job versions in place
The previous change to add build/job object version numbers in ZK
was perhaps too conservative with the way it updated them.  It
updated a copy of the version dictionary in order to avoid updating
the in-memory dictionary in case the ZK write failed.  This behavior
is common for calls to updateAttributes, but we don't get this
behavior when we use activeContext, so we generally don't rely on
it.

This change makes it slightly more efficient by updating the
in-memory dictionary in place before writing it to ZK.

Change-Id: If48d745ceefc0897946232947e5c6056fc833041