gerrit driver: fix bug around unicode branch names
If a branch name contains unicode characters that are more than 1-byte
wide, the size in bytes of the pack record won't match the size in
characters, and the pack parsing will be incorrect.

Instead, treat everything as an encoded byte string until parsing is
done - and only decode when handling a single, parsed, record.

Change-Id: I7f1a0cc96a36129fbc04c7a8687da3f66c1eef99