Dockerfile: build later skopeo
skopeo has started failing with

  unsupported MIME type for compression: application/vnd.in-toto+json

since the buildx v.10.0 release [1].  The version in bullseye is a
long way behind upstream, and unfortunately there's no practical way
to backport the package (it would require also pulling in a lot of
exploded out go container dependencies).

Thus we take the alternative solution of just building it ourselves
for the executor image.  I couldn't figure out how to build this with
the Debian-packaged go (I think it has something to do with the
aforementioned libraries) but it works with the upstream monolith.  It
produces a binary that is linked to a few local libraries, which are
installed here.

[1] https://github.com/containers/skopeo/issues/1874

Change-Id: Iab667a92a5b6e6f8591db2aa435a782913d9d34f