Filter input validation for "Change"
User sometimes tends to copy the value from the "Change" column as a input
for filtering, which would cause a 500 error in API.

To fix this we do a input validation to indicate that the change must be
a integer for example a PR bumber. So that make it clear any other fomat
like a ref or a string with "," is not supported.

Change-Id: Ia25a60429acc48c5773c808be6a900ba9bdf608e