WIP consistent patchset behavior for gerrit ssh and http
Currently the gerrit driver has different patchset specific behavior
between the ssh half of the driver and the http half. Try to make them
consistent by updating the http side to be more like ssh. The ssh half
seems to collect enough info to handle old patchsets properly, but the
http side previously would really only want to operate on the current
patchset.

Note this is incomplete and more of a memory record for what sorts of
things need to be done to aid in paging this stuff back in when I have
more time to look at it.

Change-Id: I81f240045c0b98be31c7218ce458b0a6505b5f89