dnm: Ignore change list refresh errors during pipeline processing
This change temporarily works around a race condition that can occur
after a reconfiguration while some schedulers are updating their local
layout and some already start processing pipelines in the same tenant.

In this case the pipeline manager's `_postConfig()` method that calls
`PipelineChangeList.create(...)` races with the pipeline processor
updating the change keys.

This leads to two change lists being written as separate shards, that
can't be correctly loaded, as all shards combined are expected to form a
single JSON object.

The sequence of events seems to be the following:
1. S1: pipeline processor needs to update the change keys
2. S1 the shard writer will delete the `change_list` key with the old
   shards
3. S2: configloader calls the `_postConfig()` method
4. S2: `PipelineChangeList.create()` notices that the `change_list` node
   doesn't exists in Zookeeper:
   https://opendev.org/zuul/zuul/src/branch/master/zuul/model.py#L921
6. S2: the shard writer creates the first shard `0000000000`
7. S1: the shard writer creates the second shared `0000000001`

The race condition seems to be introduced with
Ib1e467b5adb907f93bab0de61da84d2efc22e2a7

Change-Id: Iec9a9449b0457e7e8f27183ee6619e208931dabf