Address SqlAlchemy Removed in 2.0 Warnings
This addresses SqlAlchemy's removed in 2.0 warnings. Now that SqlAlchemy
2.0 has released we can see that we are not compatible yet. A good first
step in adding compatibility is fixing warnings in 1.4.

In particular there are four types of warning we fix here:

 1. Using raw strings in conn.execute() calls. We need to use the text()
    construct instead.
 2. Passing a list of items to select when doing select queries. Instead
    we need to pass things as normal posargs.
 3. Accessing row result items as if the row is a dict. THis is not
    longer possible without first going through the row._mapping system.
    Instead we can access items as normal object attributes.
 4. You must now use sqlalchemy.inspect() on a connectable to create an
    Inspector object rather than instantiating it directly.

Finally we set up alembic's engine creation to run with future 2.0
behavior now that the warnings are cleared up. This appears to have
already been done for the main zuul application.

Change-Id: I5475e39bd93d71cd1106ec6d3a5423ea2dd51859