Revert "Update git"
This reverts commit 944b9852c9f565ec433a808385d4ed0ad3f0abcc.

Upstream Debian has updated git to 2.30.2-1+deb11u1 which patches git
for the issues we manually patched to cover. We don't need the manual
patch anymore and can switch to the distro hence this revert.

Reviewers should double check that the image build process installs the
expected 2.30.2-1+deb11u1 version.

Change-Id: I5796ca327d70479931dc1c458b18d220e9ece082