Test with MariaDB instead of MySQL
I'm not sure if this is a good idea or not yet. Testing with MySQL
actually exposed the alembic transaction thing that is fixed in a
grandparent change. That said much of the world seems to be moving to
MariaDB so it might be a good idea for us to test that more explicitly?

This change has been pushed in part to collect feedback.

Change-Id: I9d00557ca5823da34278c3e9ecda2daaf13440b7