Switch our local testing docker-compose to mysql 8.0
Recently I debugged an issue that reproduced on mysql 8.0 on Jammy but
not my local system. It took quite some time for me to realize that
there was a difference in database versions which ended up being
an important detail. Update our docker-compose to better match what we
get in CI.

Change-Id: I7de268acb81680f3e6b7d3b1aa057e7babd3fa62