Drop python warnings from test suite
These were added to aid in the transition to SqlAlchemy 2.0 by calling
out issues that needed to be addressed there. This has been completed so
we drop the warnings as unnecessary configuration.

Note this might also drop warnings about kazoo using deprecated TLS
settings which may or may not be desired.

Change-Id: If6cb6e3b4db5d7db6e42a799932a1f77214a994f