Update patch set 4
Patch Set 4:

This revision addresses the previous comments and adds an attempt to initialize the objects in the postConfig call.  This is essentially an optimization to avoid spamming the logs with a very large number of harmless errors at startup (since at startup, no pipelines exist and the scheduler immediately starts trying to run cleanup routines).  This adds a small number of ZK calls to postConfig, but I think that should be negligible.  It should only attempt to initialize the objects if they don't already exist.  If they do (which is the typical case), it will continue to defer the work.

Patch-set: 4