Fix ResourceWarnings in fingergw
The fingergw (and its associated testing) was not properly managing ssl
sockets. The issue was we were in a context manager for the underlying
tcp socket which will get closed, but that doesn't call close() on the
ssl socket wrapping the tcp socket. Fix this by moving common recv()
code into a function then use the ssl socket in an inner context manager
if we are using ssl.

Both ssl and plain tcp will close() properly and we avoid duplicating
common code.

Change-Id: I1feefbd03a90734cf3c16baa6ed8f52cd8e00d14