Test job variants when override-checkout is a tag
This codifies the behavior of job variants when override-checkout
is directed at a tag.  It does indeed cause Zuul to checkout the tag
when running the job (this is covered by existing tests).  But it
does not alter the job variants which are applied.  This has the
potentially counter-intuitive effect of running a job where the
git repo on disk corresponds to one state and the configuration
of the running zuul job is a different state.

Put another way, when a user sets "override-checkout: foo", they
certainly expect zuul to checkout the tag "foo" when preparing the
git repo.  But they might also expect zuul to use the variants of
jobs that are defined in branches which contain the "foo" tag.

But since override-checkout only alters the job variant matching
behavior if the target is a branch, zuul will simply rely on the
normal behavior of matching job variants to the queue item.

These tests may be useful if we decide to update the behavior of
override-checkout.

This also makes the repr method of Ref objects more verbose, so
that it's easier to see what branches they are contained within.
This can be helpful when debugging branch matching of tags.

Change-Id: Iecd5ff269a95c89cf84783a5671fec275d492b5e