Match events to pipelines based on topic deps
We distribute tenant events to pipelines based on whether the event
matches the pipeline (ie, patchset-created for a check pipeline) or
if the event is related to a change already in the pipeline.  The
latter condition means that pipelines notice quickly when dependencies
are changed and they can take appropriate action (such as ejecting
changes which no longer have correct dependencies).

For git and commit dependencies, an update to a cycle to add a new
change requires an update to at least one existing change (for example,
adding a new change to a cycle usually requires at least two Depends-On
footers: the new change, as well as one of the changes already in the
cycle).  This means that updates to add new changes to cycles quickly
come to the attention of pipelines.

However, it is possible to add a new change to a topic dependency cycle
without updating any existing changes.  Merely uploading a new change
in the same topic adds it to the cycle.  Since that new change does
not appear in any existing pipelines, pipelines won't notice the update
until their next natural processing cycle, at which point they will
refresh dependencies of any changes they contain, and they will notice
the new dpendency and eject the cycle.

To align the behavior of topic dependencies with git and commit
dependencis, this change causes the scheduler to refresh the
dependencies of the change it is handling during tenant trigger event
processing, so that it can then compare that change's dependencies
to changes already in pipelines to determine if this event is
potentially relevant.

This moves some work from pipeline processing (which is highly parallel)
to tenant processing (which is only somewhat parallel).  This could
slow tenant event processing somewhat.  However, the work is
persisted in the change cache, and so it will not need to be repeated
during pipeline processing.

This is necessary because the tenant trigger event processor operates
only with the pipeline change list data; it does not perform a full
pipeline refresh, so it does not have access to the current queue items
and their changes in order to compare the event change's topic with
currently enqueued topics.

There are some alternative ways we could implement this if the additional
cost is an issue:

1) At the beginning of tenant trigger event processing, using the change
   list, restore each of the queue's change items from the change cache
   and compare topics.  For large queues, this could end up generating
   quite a bit of ZK traffic.
2) Add the change topic to the change reference data structure so that
   it is stored in the change list.  This is an abuse of this structure
   which otherwise exists only to store the minimum amount of information
   about a change in order to uniquely identify it.
3) Implement a PipelineTopicList similar to a PipelineChangeList for storing
   pipeline topics and accesing them without a full refresh.

Another alternative would be to accept the delayed event handling of topic
dependencies and elect not to "fix" this behavior.

Change-Id: Ia9d691fa45d4a71a1bc78cc7a4bdec206cc025c8