Re-enqueue changes if dequeued missing deps
When users create dependency cycles, the process often takes multiple
steps, some of which cause changes enqueued in earlier steps to be
dequeued.  Users then need to re-enqueue those changes in order to
have all changes in the cycle tested.

This change attempts to improve this by detecting that situation and
re-enqueing changes that are being de-queued because of missing deps.
Note, thare are plenty of cases where we de-queue because of missing
deps and we don't want to re-enqueue (ie, a one-way dependent change
ahead has failed).  To restrict this to only the situation we're
interested in, we only act if the dependencies are already in the
pipeline.  A recently updated change in a cycle will have just been
added to the pipeline, so this is true in the case we're interested in.
A one-way dependent change that failed will have already been removed
from the pipeline, and so this will be false in cases in which we
are not interested.

Change-Id: I84b3b2f8fffd1c946dafa605d1c17a37131558bd