merger: Keep redundant cherry-pick commits
In normal git usage, cherry-picking a commit that has already been
applied doesn't do anything and git exits with an error to let the user
decide what they want to do. However, this doesn't match the behavior of
merges where they will succeed even if they do nothing.

To fix this, add the --keep-redundant-commit option to `git cherry-pick`
to make git add commit when cherry-picking, even if it does nothing.

This is particularly important in a "deploy" pipeline triggered by a
Gerrit "change-merged" event, since the scheduler will try to
cherry-pick the change on top of the commit that just merged. Without
this option, the cherry-pick will fail and the deploy pipeline will fail
with a MERGE_CONFICT.

Change-Id: I326ba49e2268197662d11fd79e46f3c020675f21